Fixed adult language
This should take care of some of the adult language that's used on the blog. Feel free to add more in if you like.

I've done it in javascript rather than changed all the markdown files. That this way it's probably a bit easier to switch back if it doesn't work out.

Also, this way we could eventually have a pop up which asked if they wanted to see the adult version or not?

Interested to hear your thoughts.